# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BAL Market Inventory Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /seasons:
    get:
      operationId: listSeasons
      tags: [seasons]
      summary: List all seasons
      responses:
        "200":
          description: List of seasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"
    post:
      operationId: createSeason
      tags: [seasons]
      summary: Create a season
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeasonBody"
      responses:
        "201":
          description: Created season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/{id}:
    get:
      operationId: getSeason
      tags: [seasons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
    put:
      operationId: updateSeason
      tags: [seasons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeasonBody"
      responses:
        "200":
          description: Updated season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
    delete:
      operationId: deleteSeason
      tags: [seasons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /markets:
    get:
      operationId: listMarkets
      tags: [markets]
      summary: List all markets
      responses:
        "200":
          description: List of markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
    post:
      operationId: createMarket
      tags: [markets]
      summary: Create a market
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketBody"
      responses:
        "201":
          description: Created market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
  /markets/{id}:
    get:
      operationId: getMarket
      tags: [markets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
    put:
      operationId: updateMarket
      tags: [markets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketBody"
      responses:
        "200":
          description: Updated market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
    delete:
      operationId: deleteMarket
      tags: [markets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /storage-locations:
    get:
      operationId: listStorageLocations
      tags: [storage_locations]
      summary: List all storage locations
      responses:
        "200":
          description: List of storage locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StorageLocation"
    post:
      operationId: createStorageLocation
      tags: [storage_locations]
      summary: Create a storage location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorageLocationBody"
      responses:
        "201":
          description: Created storage location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageLocation"
  /storage-locations/{id}:
    delete:
      operationId: deleteStorageLocation
      tags: [storage_locations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Season:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        label:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, year, label, isActive, createdAt]
    CreateSeasonBody:
      type: object
      properties:
        year:
          type: integer
        label:
          type: string
        isActive:
          type: boolean
      required: [year, label]
    Market:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        region:
          type: string
        createdAt:
          type: string
      required: [id, name, country, createdAt]
    CreateMarketBody:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        region:
          type: string
      required: [name, country]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contact:
          type: string
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        contact:
          type: string
      required: [name]
    StorageLocation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateStorageLocationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
...